# docs/cs/openapi.yaml
openapi: 3.1.0
info:
  title: 2Africa CS-Core API
  version: 1.0.0-rc.1
  summary: Horizontal customer-service module shared across the 2Africa product family.
  description: |
    The CS-Core API is the read+write surface for the cross-product
    Customer Service module embedded in AgriOS / RetailOS / FactoryOS /
    TravelOS / AgriCloud / MarketOS. It is the same wire format whether
    you call it via `/v1/cs/*` (canonical) or `/v1/service/*` (legacy
    alias, kept for backward compatibility through v4.x).

    See `docs/cs/README.md` for the architectural overview and
    `docs/cs/INTEGRATION.md` for how consuming products plug in their
    own `BusinessContextProvider`.

  contact:
    name: 2Africa.AI
    url: https://github.com/2AfricaAI/2Africa-AgriOS

  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: http://localhost:8080/api
    description: Local dev (embedded in AgriOS)
  - url: http://localhost:4410/api
    description: Future standalone ServiceOS (see docs/cs/PORTS.md)
security:
  - bearerAuth: []
...
paths:
  /v1/cs/conversations:
    get:
      tags: [Conversations]
      summary: List conversations
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, pending, resolved, snoozed], default: open }
        - name: inboxId
          in: query
          schema: { type: integer, format: int64 }
        - name: assigneeType
          in: query
          schema: { type: string, enum: [me, unassigned, assigned] }
        - name: page
          in: query
          schema: { type: integer, minimum: 1, default: 1 }
      responses:
        '200':
          description: Envelope containing a page of conversation summaries.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeListConversationSummary'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/cs/conversations/{id}:
    get:
      tags: [Conversations]
      summary: Get conversation detail
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Conversation + messages + business context + WhatsApp policy block.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeConversationDetail'
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /v1/cs/conversations/{id}/messages:
    post:
      tags: [Conversations]
      summary: Reply (public outgoing or private internal note)
      description: |
        For WhatsApp inboxes, public reply is gated by Sprint 47's
        zero-cost policy: outside the 24-hour service window after a
        customer-initiated inbound, public replies fail with code
        `40901`. Private notes are always allowed.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content: { type: string, description: 'Message body the agent typed.' }
                privateNote: { type: boolean, default: false }
      responses:
        '200': { $ref: '#/components/responses/EmptyOk' }
        '400':
          description: |
            Business error (RFC 7807). Possible codes:
            - `40901` WhatsApp 24h service window expired (Sprint 47 policy)
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /v1/cs/conversations/{id}/status:
    post:
      tags: [Conversations]
      summary: Toggle conversation status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [open, resolved, pending, snoozed] }
      responses:
        '200': { $ref: '#/components/responses/EmptyOk' }
  /v1/cs/conversations/{id}/assignee:
    post:
      tags: [Conversations]
      summary: Assign or unassign an agent
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                assigneeId: { type: integer, format: int64, nullable: true, description: 'Chatwoot agent id, or null to unassign.' }
      responses:
        '200': { $ref: '#/components/responses/EmptyOk' }
...
  /v1/cs/sms-templates/render:
    post:
      tags: [Templates]
      summary: Render a template against a conversation's customer context
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code: { type: string }
                conversationId: { type: integer, format: int64 }
      responses:
        '200':
          description: Rendered body envelope.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvelopeRenderedTemplate'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    NotFound:
      description: Resource not found.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    InternalError:
      description: Server error.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    EmptyOk:
      description: Empty OK envelope.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeEmpty' }
  schemas:
...
    EnvelopeMeta:
      type: object
      properties:
        code: { type: integer, example: 200 }
        msg: { type: string, example: OK }
        traceId: { type: string, format: uuid }
        timestamp: { type: string, format: date-time }
      required: [code, msg]
    EnvelopeEmpty:
      allOf:
        - $ref: '#/components/schemas/EnvelopeMeta'
        - type: object
          properties:
            data: { type: 'null' }
    EnvelopeListConversationSummary:
      allOf:
        - $ref: '#/components/schemas/EnvelopeMeta'
        - type: object
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ConversationSummary' }
    EnvelopeConversationDetail:
      allOf:
        - $ref: '#/components/schemas/EnvelopeMeta'
        - type: object
          properties:
            data: { $ref: '#/components/schemas/ConversationDetail' }
...
    EnvelopeRenderedTemplate:
      allOf:
        - $ref: '#/components/schemas/EnvelopeMeta'
        - type: object
          properties:
            data:
              type: object
              properties:
                rendered: { type: string }
                code: { type: string }
                customerId: { type: integer, format: int64, nullable: true }
...
    ConversationSummary:
      type: object
      properties:
        id: { type: integer, format: int64 }
        displayId: { type: integer, format: int64 }
        status: { type: string, enum: [open, pending, resolved, snoozed] }
        channel: { type: string, description: 'Chatwoot channel_type, e.g. Channel::WebWidget' }
        inboxId: { type: integer, format: int64 }
        subjectId: { type: integer, format: int64, nullable: true, description: 'cs_contact_link.subject_id when linked' }
        subjectType: { type: string, nullable: true, description: 'customer / buyer / traveler / etc.' }
        contactName: { type: string, nullable: true }
        contactPhone: { type: string, nullable: true }
        contactEmail: { type: string, nullable: true, format: email }
        lastMessagePreview: { type: string, nullable: true }
        lastActivityAt: { type: integer, format: int64, description: 'Unix epoch seconds' }
        unreadCount: { type: integer, default: 0 }
        assigneeId: { type: integer, format: int64, nullable: true }
    ConversationDetail:
      allOf:
        - $ref: '#/components/schemas/ConversationSummary'
        - type: object
          properties:
            messages:
              type: array
              items: { $ref: '#/components/schemas/Message' }
            businessContext: { $ref: '#/components/schemas/BusinessContext' }
            whatsAppPolicy: { $ref: '#/components/schemas/WhatsAppPolicy' }
    Message:
      type: object
      properties:
        id: { type: integer, format: int64 }
        content: { type: string }
        messageType: { type: integer, enum: [0, 1], description: '0 = incoming customer, 1 = outgoing agent/bot' }
        privateNote: { type: boolean }
        createdAt: { type: integer, format: int64, description: 'Unix epoch seconds' }
    BusinessContext:
      type: object
      description: Product-specific aggregation produced by `BusinessContextProvider` (SPI).
      properties:
        openOrderCount: { type: integer, default: 0 }
        overdueArInvoiceCount: { type: integer, default: 0, nullable: true }
        overdueArAmount: { type: number, nullable: true }
        openComplaintCount: { type: integer, default: 0, nullable: true }
        lastOrderDate: { type: string, format: date, nullable: true }
    WhatsAppPolicy:
      type: object
      description: Sprint 47 service-window policy snapshot for this conversation.
      properties:
        managed: { type: boolean, description: 'True when this inbox is governed by the WhatsApp policy.' }
        lastInboundAt: { type: integer, format: int64, nullable: true }
        serviceWindowExpiresAt: { type: integer, format: int64, nullable: true }
        withinServiceWindow: { type: boolean }
...
    Problem:
      type: object
      description: RFC 7807-style error envelope. CS-Core also returns the legacy AgriOS `R<T>` shape for backward compatibility; new callers should treat 4xx as `problem+json`.
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer, minimum: 100, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri, nullable: true }
        code: { type: integer, description: 'Business code (e.g. 40901 = WhatsApp window expired). Aligns with AgriOS R<T>.code.' }
        traceId: { type: string, format: uuid }